Moving junit tests
