+++ console/src/test/java/org/eclipse/rdf4j/console/ConsoleIOTest.java
@@ -14,6 +14,8 @@ import static org.mockito.Mockito.mock;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
+import org.eclipse.rdf4j.console.ConsoleIO;
+import org.eclipse.rdf4j.console.ConsoleState;
 
 import org.junit.Before;
 import org.junit.Test;
