+++ console/src/test/java/org/eclipse/rdf4j/console/command/DropTest.java
@@ -5,13 +5,11 @@
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/org/documents/edl-v10.php.
  *******************************************************************************/
-package org.eclipse.rdf4j.console;
+package org.eclipse.rdf4j.console.command;
 
-import org.eclipse.rdf4j.console.command.Close;
-import org.eclipse.rdf4j.console.command.Drop;
 import static org.hamcrest.core.Is.is;
 import static org.hamcrest.core.IsEqual.equalTo;
-import static org.junit.Assert.assertThat;
+
 import static org.mockito.Matchers.anyBoolean;
 import static org.mockito.Matchers.startsWith;
 import static org.mockito.Mockito.mock;
@@ -23,10 +21,14 @@ import java.io.IOException;
 import java.io.UnsupportedEncodingException;
 
 import org.eclipse.rdf4j.RDF4JException;
+import org.eclipse.rdf4j.console.ConsoleState;
 import org.eclipse.rdf4j.repository.RepositoryException;
 import org.eclipse.rdf4j.repository.config.RepositoryConfig;
 import org.eclipse.rdf4j.repository.manager.LocalRepositoryManager;
 import org.eclipse.rdf4j.repository.sail.config.ProxyRepositoryConfig;
+
+import static org.junit.Assert.assertThat;
+
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Rule;
